Moved and renamed files to reorganise project
